Enable drag & drop to an empty section (requires a section footer or header)
In the original implementation it's impossible to drag & drop a cell into an empty section since the logic looks for an adjacent cell to derive an NSIndexPath. Because an empty section does not have any cells the logic looks for the CGRect of the header or footer while dragging the table cell only if it doesn't have a current match for a table cell. When there is a hit for a header or footer the next NSIndexPath is assumed to be section #, row 0 in the section that lacks any table cells.